requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cloudinary==1.24.0
dj-database-url==0.5.0
Django==3.2a1
django-ckeditor==6.0.0
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow==8.1.0
psycopg2-binary==2.8.6
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
